# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateSeatGridRequest:
    description: Seat grid creation request data
    properties:
      category:
        allOf:
        - $ref: '#/definitions/models.SeatCategory'
        enum:
        - standard
        - vip
        - accessible
        - premium
        - economy
        - standing
        example: standard
      end_number:
        example: 20
        minimum: 1
        type: integer
      end_row:
        example: Z
        maxLength: 10
        minLength: 1
        type: string
      spacing_x:
        example: 1.2
        minimum: 0.1
        type: number
      spacing_y:
        example: 1.2
        minimum: 0.1
        type: number
      start_number:
        example: 1
        minimum: 1
        type: integer
      start_row:
        example: A
        maxLength: 10
        minLength: 1
        type: string
      start_x:
        example: 0
        type: number
      start_y:
        example: 0
        type: number
    required:
    - category
    - end_number
    - end_row
    - spacing_x
    - spacing_y
    - start_number
    - start_row
    - start_x
    - start_y
    type: object
  models.CreateSeatRequest:
    description: Seat creation request data
    properties:
      category:
        allOf:
        - $ref: '#/definitions/models.SeatCategory'
        enum:
        - standard
        - vip
        - accessible
        - premium
        - economy
        - standing
        example: standard
      height:
        example: 1
        minimum: 0.1
        type: number
      number:
        example: "1"
        maxLength: 10
        minLength: 1
        type: string
      rotation:
        example: 0
        type: number
      row:
        example: A
        maxLength: 10
        minLength: 1
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.SeatStatus'
        enum:
        - available
        - occupied
        - reserved
        - blocked
        - maintenance
        example: available
      width:
        example: 1
        minimum: 0.1
        type: number
      x:
        example: 10.5
        type: number
      "y":
        example: 15.2
        type: number
    required:
    - category
    - number
    - row
    - status
    - x
    - "y"
    type: object
  models.CreateSeatingAssignmentRequest:
    description: Seating assignment creation request data
    properties:
      guest_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      notes:
        example: VIP guest - front row preference
        maxLength: 500
        type: string
      seat_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - guest_id
    - seat_id
    type: object
...
  models.Event:
    description: Event information and details
    properties:
      allow_self_rsvp:
        description: Guest Management Options
        type: boolean
      categories:
        description: Discovery and Organization (for public events)
        items:
          type: string
        type: array
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      date:
        example: "2024-06-15T18:00:00Z"
        type: string
      description:
        example: A beautiful celebration of love
        type: string
      guests:
        description: Guest Management
        items:
          $ref: '#/definitions/models.Guest'
        type: array
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      location:
        example: Grand Hotel Ballroom
        type: string
      max_guests:
        type: integer
      name:
        example: John & Jane's Wedding
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: string
      require_approval:
        type: boolean
      slug:
        example: john-jane-wedding-2024
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      visibility:
        allOf:
        - $ref: '#/definitions/models.EventVisibility'
        description: Visibility and Access Control
    type: object
...
  models.EventVisibility:
    enum:
    - private
    - public
    type: string
    x-enum-varnames:
    - EventVisibilityPrivate
    - EventVisibilityPublic
...
  models.Guest:
    description: Guest information and RSVP status
    properties:
      approved:
        type: boolean
      created_at:
        description: Timestamps
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        description: Guest Information
        example: John Doe
        type: string
      notes:
        example: Vegetarian, allergic to nuts
        type: string
      phone:
        example: "+1234567890"
        type: string
      rsvp_date:
        example: "2024-01-15T10:30:00Z"
        type: string
      rsvp_status:
        allOf:
        - $ref: '#/definitions/models.RSVPStatus'
        description: RSVP Status
      seat:
        $ref: '#/definitions/models.Seat'
      seat_id:
        description: Seating (for future implementation)
        type: string
      source:
        allOf:
        - $ref: '#/definitions/models.GuestSource'
        description: Source and Approval
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        description: User relationship (only if guest is a platform user) - NEW FIELD
        type: string
    type: object
...
  models.RSVPStatus:
    enum:
    - pending
    - accept
    - decline
    - maybe
    type: string
    x-enum-varnames:
    - RSVPStatusPending
    - RSVPStatusAccept
    - RSVPStatusDecline
    - RSVPStatusMaybe
...
  models.Seat:
    description: Seat information and details
    properties:
      category:
        $ref: '#/definitions/models.SeatCategory'
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      guest:
        $ref: '#/definitions/models.Guest'
      guest_id:
        type: string
      height:
        example: 1
        type: number
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      number:
        example: "1"
        type: string
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: string
      rotation:
        example: 0
        type: number
      row:
        example: A
        type: string
      status:
        $ref: '#/definitions/models.SeatStatus'
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      width:
        example: 1
        type: number
      x:
        description: Position X coordinate
        example: 10.5
        type: number
      "y":
        description: Position Y coordinate
        example: 15.2
        type: number
    type: object
  models.SeatCategory:
    enum:
    - standard
    - vip
    - accessible
    - premium
    - economy
    - standing
    type: string
    x-enum-varnames:
    - SeatCategoryStandard
    - SeatCategoryVIP
    - SeatCategoryAccessible
    - SeatCategoryPremium
    - SeatCategoryEconomy
    - SeatCategoryStanding
  models.SeatResponse:
    description: Seat response data
    properties:
      seat:
        $ref: '#/definitions/models.Seat'
    type: object
  models.SeatStatus:
    enum:
    - available
    - occupied
    - reserved
    - blocked
    - maintenance
    type: string
    x-enum-varnames:
    - SeatStatusAvailable
    - SeatStatusOccupied
    - SeatStatusReserved
    - SeatStatusBlocked
    - SeatStatusMaintenance
  models.SeatingAssignment:
    description: Seating assignment information and details
    properties:
      assigned_at:
        type: string
      assigned_by:
        type: string
      assigned_by_user:
        $ref: '#/definitions/models.User'
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      guest:
        $ref: '#/definitions/models.Guest'
      guest_id:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      notes:
        example: VIP guest - front row preference
        type: string
      seat:
        $ref: '#/definitions/models.Seat'
      seat_id:
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  models.SeatingAssignmentResponse:
    description: Seating assignment response data
    properties:
      seating_assignment:
        $ref: '#/definitions/models.SeatingAssignment'
    type: object
  models.SeatingAssignmentsResponse:
    description: Multiple seating assignments response data
    properties:
      seating_assignments:
        items:
          $ref: '#/definitions/models.SeatingAssignment'
        type: array
      total:
        type: integer
    type: object
  models.SeatsResponse:
    description: Multiple seats response data
    properties:
      seats:
        items:
          $ref: '#/definitions/models.Seat'
        type: array
      total:
        type: integer
    type: object
...
  models.UpdateSeatRequest:
    description: Seat update request data
    properties:
      category:
        allOf:
        - $ref: '#/definitions/models.SeatCategory'
        enum:
        - standard
        - vip
        - accessible
        - premium
        - economy
        - standing
        example: vip
      height:
        example: 1.2
        minimum: 0.1
        type: number
      number:
        example: "2"
        maxLength: 10
        minLength: 1
        type: string
      rotation:
        example: 45
        type: number
      row:
        example: B
        maxLength: 10
        minLength: 1
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.SeatStatus'
        enum:
        - available
        - occupied
        - reserved
        - blocked
        - maintenance
        example: reserved
      width:
        example: 1.2
        minimum: 0.1
        type: number
      x:
        example: 12
        type: number
      "y":
        example: 18
        type: number
    type: object
...
  models.User:
    description: User account information
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_name:
        example: Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
      profile_pic:
        example: https://example.com/profile.jpg
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the SeatMaster backend server for event management and seating
    arrangements.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SeatMaster Backend API
  version: "1.0"
paths:
...
  /events/{id}/seating:
    get:
      description: Retrieve all seat assignments for an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeatingAssignmentsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get seat assignments
      tags:
      - seating
  /events/{id}/seating/assign:
    post:
      consumes:
      - application/json
      description: Assign a guest to a specific seat for an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seatId
        required: true
        type: string
      - description: Seating assignment data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/models.CreateSeatingAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SeatingAssignmentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign guest to seat
      tags:
      - seating
  /events/{id}/seating/assign/{seatId}:
    delete:
      description: Remove a guest from a specific seat
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unassign guest from seat
      tags:
      - seating
...
  /venues/{id}/rooms/{roomId}/seats:
    get:
      description: Retrieve all seats in a specific room
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get seats in room
      tags:
      - seats
    post:
      consumes:
      - application/json
      description: Create a new seat in a room
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Seat creation data
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/models.CreateSeatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SeatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new seat
      tags:
      - seats
  /venues/{id}/rooms/{roomId}/seats/{seatId}:
    delete:
      description: Delete an existing seat
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seatId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete seat
      tags:
      - seats
    patch:
      consumes:
      - application/json
      description: Update an existing seat
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seatId
        required: true
        type: string
      - description: Seat update data
        in: body
        name: seat
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSeatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update seat
      tags:
      - seats
  /venues/{id}/rooms/{roomId}/seats/grid:
    post:
      consumes:
      - application/json
      description: Create multiple seats in a grid layout
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Seat grid creation data
        in: body
        name: grid
        required: true
        schema:
          $ref: '#/definitions/models.CreateSeatGridRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SeatsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a grid of seats
      tags:
      - seats
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"